refactor(imports): update component imports to use Atoms index
